third_party/WebKit/LayoutTests/imported/wpt/html/browsers/browsing-the-web/navigating-across-documents/012-expected.txt

CONSOLE WARNING: line 11: Synchronous XMLHttpRequest on the main thread is deprecated because of its detrimental effects to the end user's experience. For more help, check https://xhr.spec.whatwg.org/.
This is a testharness.js-based test.
FAIL Link with onclick navigation to javascript url with delayed document.write and href navigation  assert_equals: expected "href" but got "write"
Harness: the test ran to completion.
